Re-download files when size mismatch

+++ astro_plasma/core/utils.py
@@ -27,7 +27,7 @@ load_dotenv()
 
 # download chunk size, # bytes to download at a time
 CHUNK_SIZE = int(os.getenv("CHUNK_SIZE", "4096"))
-BASE_URL = os.getenv("WEB_SERVER_BASE_URL", "http://localhost:8000")
+BASE_URL = os.getenv("ASTROPLASMA_SERVER", "http://localhost:8000")
 PARALLEL_JOBS = int(os.getenv("PARALLEL_DOWNLOAD_JOBS", "3"))
 
 
@@ -107,20 +107,14 @@ def parse_atomic_ion_no(
 
 
 def fetch(urls: List[Tuple[str, Path]], base_dir: Path):
-    if not base_dir.exists():
-        base_dir.mkdir(mode=0o766, parents=True)
-
-    # extract urls and file paths iff it not exists
-    downlodable_urls = []
-    for url, file_name in urls:
-        save_path = base_dir / file_name
-        if not save_path.exists():
-            downlodable_urls.append((url, save_path))
+    base_dir.mkdir(mode=0o766, parents=True, exist_ok=True)
 
     # unit job of downloading and saving file
-    def download_job(url, save_path):
+    def download_job(url: str, save_path: Path):
+        print(url)
         response = requests.get(url, stream=True)
         file_size = int(response.headers.get("content-length", 0))
+
         progress = tqdm(
             desc=save_path.name,
             total=file_size,
@@ -129,6 +123,10 @@ def fetch(urls: List[Tuple[str, Path]], base_dir: Path):
             unit="iB",
             leave=False,
         )
+
+        if save_path.exists() and save_path.stat().st_size == file_size:
+            progress.update(file_size)
+            return
 
         with open(save_path, "wb") as file:
             for data in response.iter_content(chunk_size=CHUNK_SIZE):
@@ -139,6 +137,6 @@ def fetch(urls: List[Tuple[str, Path]], base_dir: Path):
 
     with ThreadPoolExecutor(max_workers=PARALLEL_JOBS, thread_name_prefix="cloudy_dnldr") as pool:
         # iterative download and show progress bar using tqdm
-        for url, save_path in downlodable_urls:
-            url = urljoin(base=BASE_URL, url=url)
-            pool.submit(download_job, url, save_path)
+        for url_path, file_name in urls:
+            url = urljoin(BASE_URL, url_path)
+            pool.submit(download_job, url, base_dir / file_name)
